Create FileZipper protocol
